Фильтр для даты остался
#10
