+++ front/src/layouts/Header.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useNavigate } from 'react-router-dom';
 
-import { currentUser } from '../packages/API/bd';
+import { currentUser } from '../Components/bd';
 import CompanyLogo from '../Components/templateStyle/CompanyLogo'
 import NavLinks from './NavLinks';
 import avatar from '../Assets/img/Accountant.png'
