--- front/src/packages/API/bd.js
@@ -1,268 +1,0 @@
-import Button from "../../Components/Button";
-import check from '../../Assets/img/check.png'
-
-export const heaaders = {
-    number: "Номер заказа",
-    address: "адресс",
-    time: "Доставка к",
-    summ: "сумма",
-    size: "Габариты",
-    pay: "Оплачено",
-    Status: "Состояние",
-    Open: "Открыть",
-
-};
-
-export const data = [
-    {
-        id: '19',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '18',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '17',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '16',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '15',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '14',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '13',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '12',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '11',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '10',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '9',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '8',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '7',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '6',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '5',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '94',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '93',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.success + 'button__fs20'}>Отправлено</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '91',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.warning + 'button__fs20'}>На складе</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-    {
-        id: '92',
-        number: "E-909",
-        address: "Торжковская, 15",
-        time: <div>22.10.2022 <br/><br/>14:00-15:00</div>,
-        summ: "5000",
-        size: "20x30x10",
-        pay: <img src={check}/>,
-        Status: <Button className={Button.style.danger + 'button__fs20'}>Не прибыло</Button>,
-        Open: <Button className={Button.style.success + 'button__fs20'}>Открыть</Button>,
-    },
-];
-
-
-export const headersShifts = {
-    fullName: "ФИО",
-    position: "Должность",
-    schedule: "График"
-};
-
-export const dataShifts = [
-    {
-        id: '1',
-        fullName: "Иерусалимов Никита",
-        position: "Администратор",
-        schedule: <div>пн-пт <br/><br/>8:00 - 18:00</div>,
-       
-    },{
-        id: '2',
-        fullName: "Сорокумов Сергей",
-        position: "Администратор",
-        schedule: <div>пн-пт <br/><br/>8:00 - 18:00</div>,
-       
-    },
-    {
-        id: '3',
-        fullName: "Голубева Валентина",
-        position: "Администратор",
-        schedule: <div>пн-пт <br/><br/>8:00 - 18:00</div>,
-       
-    },
-    {
-        id: '4',
-        fullName: "Иерусалимов Никита",
-        position: "Администратор",
-        schedule: <div>пн-пт <br/><br/>8:00 - 18:00</div>,
-       
-    },
-]
-
-export const currentUser = {
-    name:"Maria",
-    position:"accountant"
-}
